refactor(dashboard): wrap isRedirectError in shared rethrowRedirect helper
Centralizes the undocumented next/dist/client/components/redirect-error
import into lib/next/rethrow-redirect.ts. All 5 dashboard content
components and their tests now import from the shared helper.

Closes #504
