+++ apps/web/app/app/admin/dashboard/_components/weak-topics-content.test.tsx
@@ -4,7 +4,7 @@ import { beforeEach, describe, expect, it, vi } from 'vitest'
 // ---- Hoisted mocks ----------------------------------------------------------
 
 const mockGetWeakTopics = vi.hoisted(() => vi.fn())
-const mockIsRedirectError = vi.hoisted(() => vi.fn())
+const mockRethrowRedirect = vi.hoisted(() => vi.fn())
 
 // ---- Module mocks -----------------------------------------------------------
 
@@ -12,8 +12,8 @@ vi.mock('../queries', () => ({
   getWeakTopics: mockGetWeakTopics,
 }))
 
-vi.mock('next/dist/client/components/redirect-error', () => ({
-  isRedirectError: mockIsRedirectError,
+vi.mock('@/lib/next/rethrow-redirect', () => ({
+  rethrowRedirect: mockRethrowRedirect,
 }))
 
 vi.mock('./weak-topics-list', () => ({
@@ -29,7 +29,6 @@ import { WeakTopicsContent } from './weak-topics-content'
 describe('WeakTopicsContent', () => {
   beforeEach(() => {
     vi.resetAllMocks()
-    mockIsRedirectError.mockReturnValue(false)
   })
 
   it('renders the weak topics list when data loads successfully', async () => {
@@ -55,7 +54,9 @@ describe('WeakTopicsContent', () => {
   it('re-throws redirect errors instead of showing the fallback', async () => {
     const redirectError = new Error('NEXT_REDIRECT:/auth/login')
     mockGetWeakTopics.mockRejectedValue(redirectError)
-    mockIsRedirectError.mockReturnValue(true)
+    mockRethrowRedirect.mockImplementation((err: unknown) => {
+      throw err
+    })
 
     await expect(WeakTopicsContent()).rejects.toThrow('NEXT_REDIRECT:/auth/login')
   })
