+++ apps/web/app/app/admin/dashboard/_components/weak-topics-content.tsx
@@ -1,4 +1,4 @@
-import { isRedirectError } from 'next/dist/client/components/redirect-error'
+import { rethrowRedirect } from '@/lib/next/rethrow-redirect'
 import { getWeakTopics } from '../queries'
 import { ContentErrorFallback } from './content-error-fallback'
 import { WeakTopicsList } from './weak-topics-list'
@@ -8,7 +8,7 @@ export async function WeakTopicsContent() {
     const topics = await getWeakTopics()
     return <WeakTopicsList topics={topics} />
   } catch (error) {
-    if (isRedirectError(error)) throw error
+    rethrowRedirect(error)
     return <ContentErrorFallback message="Failed to load weak topics. Please refresh the page." />
   }
 }
